Use libdir instead of prefix/lib.
        * libgda/Makefile.am:
        * providers/firebird/Makefile.am:
        * providers/freetds/Makefile.am:
        * providers/ibmdb2/Makefile.am:
        * providers/ldap/Makefile.am:
        * providers/mdb/Makefile.am:
        * providers/msql/Makefile.am:
        * providers/mysql/Makefile.am:
        * providers/mysql/gda-mysql-provider.c:
        * providers/odbc/Makefile.am:
        * providers/oracle/Makefile.am:
        * providers/postgres/Makefile.am:
        * providers/sqlite/Makefile.am:
        * providers/sybase/Makefile.am:
        * providers/xbase/Makefile.am:
        * providers/xml/Makefile.am:
        Use libdir instead of prefix/lib.

        * libgda/gda-config.c: (get_config_client):
        * tools/gda-config-tool.c: (config_load_file):
        Use correct type (fix for 64bits platform).

        * providers/mysql/gda-mysql-provider.c:
        Add missing header.

+++ libgda/gda-config.c
@@ -221,7 +221,7 @@ get_config_client ()
 	 *	  wait until the operation finishes
 	 */
 	if (config_client == NULL){
-		gint len;
+		gsize  len;
 		gchar *full_file;
 		gchar *user_config;
